chore:Cleanups, reorganizings
this entire codebase man i swear.....
