+++ app/src/main/java/is/xyz/mpv/MPVActivity.kt
@@ -51,7 +51,16 @@ import com.google.android.material.bottomsheet.BottomSheetBehavior
 import com.google.android.material.bottomsheet.BottomSheetDialog
 import com.google.android.material.color.DynamicColors
 import com.google.android.material.dialog.MaterialAlertDialogBuilder
+import `is`.xyz.filepicker.FilePickerActivity
+import `is`.xyz.mpv.controls.DecimalPickerDialog
+import `is`.xyz.mpv.controls.PickerDialog
+import `is`.xyz.mpv.controls.PlaylistDialog
+import `is`.xyz.mpv.controls.SliderPickerDialog
+import `is`.xyz.mpv.controls.SubDelayDialog
+import `is`.xyz.mpv.controls.SubTrackDialog
 import `is`.xyz.mpv.databinding.PlayerBinding
+import `is`.xyz.mpv.services.BackgroundPlaybackService
+import `is`.xyz.mpv.services.NotificationButtonReceiver
 import java.io.File
 import kotlin.math.roundToInt
