+++ app/src/main/java/is/xyz/mpv/controls/PickerDialog.kt
@@ -1,4 +1,4 @@
-package `is`.xyz.mpv
+package `is`.xyz.mpv.controls
 
 import android.view.LayoutInflater
 import android.view.View
