+++ app/src/main/java/is/xyz/mpv/controls/SubDelayDialog.kt
@@ -1,4 +1,4 @@
-package `is`.xyz.mpv
+package `is`.xyz.mpv.controls
 
 import `is`.xyz.mpv.databinding.DialogDecimalBinding
 import android.text.Editable
@@ -6,7 +6,6 @@ import android.text.TextWatcher
 import android.view.LayoutInflater
 import android.view.View
 import android.widget.EditText
-import android.widget.TextView
 import androidx.core.view.isVisible
 
 internal class SubDelayDialog(
